chore: add retrieval routine for special characters
This subroutine will add a special treatment of special characters passed in the retrieval engine of the angular component. The pattern, I have chosen is the strategy pattern which is a behavioral design pattern that lets you define a family of algorithms, put each of them into a separate class, and make their objects interchangeable.

In this case, I've decided that we will make the engine pluggable with more layers to come in the future. Like for right now, I cannot think of any further use cases and the project will be terminated in 2 weeks but the usage of patterns is essential to prove that you are the smartest guy in the room. Go read about it after ordering Design Patterns: Elements of Reusable Software from your local bookstore (dont use amazon ffs)